requirements.txt
aiofiles
aiogram
bcrypt
Brotli
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
et-xmlfile
fastapi==0.63.0
Flask
gevent
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
greenlet
gspread
gspread-dataframe
h11
hdbcli
httplib2
idna
itsdangerous
Jinja2==3.0.3
MarkupSafe
numpy
oauth2client
oauthlib
opencv-python
openpyxl
pandas
paramiko
Pillow
protobuf
psycopg2
pyasn1
pyasn1-modules
pycparser
pydantic
pygsheets
pyhdb
PyNaCl
pyparsing
pysftp
python-dateutil
python-multipart
pytz
pyzmq
qrcode
requests
requests-oauthlib
roundrobin
rsa
six
SQLAlchemy
sqlalchemy-hana
starlette
transliterate
typing-extensions
uritemplate
urllib3
uvicorn
Werkzeug
zope.event
zope.interface
